requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
python-jose>=0.6.0
passlib[bcrypt]>=1.7.4
requests>=2.30.0
pytest>=7.3.1
fastapi-testclient>=0.2.0
pylint>=2.17.4